+++ src/main/java/top/supcar/server/update/WorldUpdater.java
@@ -29,7 +29,7 @@ import java.time.Instant;
 public class WorldUpdater {
 	private double timeQuant = 0.01;
 	private SessionObjects sessionObjects;
-	private int X = 1;
+	private double X = 1;
 	public static final double FIRST_QUANT = 0.01;
 
 
@@ -42,9 +42,9 @@ public class WorldUpdater {
 	public void update() {
 		Instant instant = Instant.now();
 		sessionObjects.setCurrInstant(instant);
-		long timeQuantMillis = (long)(1000*FIRST_QUANT)*X;
+		long timeQuantMillis = (long)((1000*FIRST_QUANT)*X);
 		if(lastInstant != null)
-			timeQuantMillis = Duration.between(lastInstant,instant).toMillis()*X;
+			timeQuantMillis = (long)(Duration.between(lastInstant,instant).toMillis()*X);
 		timeQuant = ((double)timeQuantMillis)/1000;
 		lastInstant = instant;
 		sessionObjects.getCarsUpdater().update();
@@ -55,7 +55,7 @@ public class WorldUpdater {
 		return timeQuant;
 	}
 
-	public void setX (int X) {
+	public void setX (double X) {
 	    this.X = X;
         System.out.println("X: " + X);
     }
